linkedin_extractor/requirements.txt
playwright
beautifulsoup4
pandas
python-dotenv
jinja2
